Moved starting point of the server to main.go and moved server to the server package Recreate go.mod to fix build issues Add new package folder and files And minor fixes
