+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"tinywebserver/server"
+)
+
+func main() {
+	server.Start()
+}
