+++ server/tiny_webserver_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"bytes"
